📝 Refactoring (utils package)
