+++ src/main/java/io/github/syst3ms/skriptparser/Parser.java
@@ -6,8 +6,8 @@ import io.github.syst3ms.skriptparser.parsing.ScriptLoader;
 import io.github.syst3ms.skriptparser.registration.DefaultRegistration;
 import io.github.syst3ms.skriptparser.registration.SkriptAddon;
 import io.github.syst3ms.skriptparser.registration.SkriptRegistration;
-import io.github.syst3ms.skriptparser.util.ConsoleColors;
 import io.github.syst3ms.skriptparser.util.FileUtils;
+import io.github.syst3ms.skriptparser.util.color.ConsoleColors;
 
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
