+++ src/main/java/io/github/syst3ms/skriptparser/expressions/CondExprContains.java
@@ -5,8 +5,8 @@ import io.github.syst3ms.skriptparser.lang.Expression;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.lang.base.ConditionalExpression;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
-import io.github.syst3ms.skriptparser.util.CollectionUtils;
 
+import java.util.Arrays;
 import java.util.Optional;
 
 /**
@@ -32,13 +32,13 @@ public class CondExprContains extends ConditionalExpression {
         );
     }
 
-    private Expression<?> first, second;
+    private Expression<?> haystack, needle;
     private boolean onlyString;
 
     @Override
     public boolean init(Expression<?>[] expressions, int matchedPattern, ParseContext parseContext) {
-        first = expressions[0];
-        second = expressions[1];
+        haystack = expressions[0];
+        needle = expressions[1];
         onlyString = matchedPattern == 0;
         setNegated(parseContext.getParseMark() == 1);
         return true;
@@ -48,18 +48,18 @@ public class CondExprContains extends ConditionalExpression {
     @Override
     public boolean check(TriggerContext ctx) {
         if (onlyString) {
-            Optional<? extends String> f = ((Expression<String>) first).getSingle(ctx);
-            Optional<? extends String> s = ((Expression<String>) second).getSingle(ctx);
-            return isNegated() != f.filter(o1 -> s.map(o1::contains).isPresent()).isPresent();
+            Optional<? extends String> haystackValue = ((Expression<String>) haystack).getSingle(ctx);
+            Optional<? extends String> needleValue = ((Expression<String>) needle).getSingle(ctx);
+            return isNegated() != haystackValue.filter(o1 -> needleValue.map(o1::contains).isPresent()).isPresent();
         } else {
-            Object[] f = ((Expression<Object>) first).getValues(ctx);
-            Object[] s = ((Expression<Object>) second).getValues(ctx);
-            return isNegated() != CollectionUtils.contains(f, s);
+            Object[] haystackValues = ((Expression<Object>) haystack).getValues(ctx);
+            Object[] needleValues = ((Expression<Object>) needle).getValues(ctx);
+            return isNegated() != Arrays.asList(haystackValues).containsAll(Arrays.asList(needleValues));
         }
     }
 
     @Override
     public String toString(TriggerContext ctx, boolean debug) {
-        return first.toString(ctx, debug) + (isNegated() ? " does not contain " : " contains ") + second.toString(ctx, debug);
+        return haystack.toString(ctx, debug) + (isNegated() ? " does not contain " : " contains ") + needle.toString(ctx, debug);
     }
 }
