+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprDateFormatted.java
@@ -4,8 +4,8 @@ import io.github.syst3ms.skriptparser.Parser;
 import io.github.syst3ms.skriptparser.lang.Expression;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
-import io.github.syst3ms.skriptparser.util.DoubleOptional;
-import io.github.syst3ms.skriptparser.util.SkriptDate;
+import io.github.syst3ms.skriptparser.util.classes.DoubleOptional;
+import io.github.syst3ms.skriptparser.util.classes.SkriptDate;
 
 /**
  * Formats a date as a string using the given format.
