+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprDateTodayAt.java
@@ -4,8 +4,8 @@ import io.github.syst3ms.skriptparser.Parser;
 import io.github.syst3ms.skriptparser.lang.Expression;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
-import io.github.syst3ms.skriptparser.util.SkriptDate;
-import io.github.syst3ms.skriptparser.util.Time;
+import io.github.syst3ms.skriptparser.util.classes.SkriptDate;
+import io.github.syst3ms.skriptparser.util.classes.Time;
 
 import java.time.Duration;
