+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprLoopValue.java
@@ -11,7 +11,7 @@ import io.github.syst3ms.skriptparser.sections.SecLoop;
 import io.github.syst3ms.skriptparser.types.TypeManager;
 import io.github.syst3ms.skriptparser.types.conversions.Converters;
 import io.github.syst3ms.skriptparser.util.ClassUtils;
-import io.github.syst3ms.skriptparser.util.Pair;
+import io.github.syst3ms.skriptparser.util.classes.Pair;
 import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.Array;
