+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprTimeInformation.java
@@ -5,7 +5,7 @@ import io.github.syst3ms.skriptparser.lang.Expression;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.lang.properties.PropertyExpression;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
-import io.github.syst3ms.skriptparser.util.Time;
+import io.github.syst3ms.skriptparser.util.classes.Time;
 
 import java.math.BigInteger;
