+++ src/main/java/io/github/syst3ms/skriptparser/sections/SecMap.java
@@ -13,7 +13,7 @@ import io.github.syst3ms.skriptparser.log.SkriptLogger;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
 import io.github.syst3ms.skriptparser.parsing.ParserState;
 import io.github.syst3ms.skriptparser.types.changers.ChangeMode;
-import io.github.syst3ms.skriptparser.util.Pair;
+import io.github.syst3ms.skriptparser.util.classes.Pair;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
