+++ src/main/java/io/github/syst3ms/skriptparser/util/classes/DoubleOptional.java
@@ -1,4 +1,4 @@
-package io.github.syst3ms.skriptparser.util;
+package io.github.syst3ms.skriptparser.util.classes;
 
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.Nullable;
