+++ src/main/java/io/github/syst3ms/skriptparser/util/classes/MultiMap.java
@@ -1,4 +1,4 @@
-package io.github.syst3ms.skriptparser.util;
+package io.github.syst3ms.skriptparser.util.classes;
 
 import org.jetbrains.annotations.Nullable;
