+++ src/main/java/io/github/syst3ms/skriptparser/util/classes/Pair.java
@@ -1,4 +1,4 @@
-package io.github.syst3ms.skriptparser.util;
+package io.github.syst3ms.skriptparser.util.classes;
 
 /**
  * A simple pair of two values.
