+++ src/main/java/io/github/syst3ms/skriptparser/util/classes/Time.java
@@ -1,4 +1,4 @@
-package io.github.syst3ms.skriptparser.util;
+package io.github.syst3ms.skriptparser.util.classes;
 
 import java.time.Duration;
 import java.time.LocalTime;
