+++ src/main/java/io/github/syst3ms/skriptparser/util/classes/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package io.github.syst3ms.skriptparser.util.classes;
+
+import javax.annotation.ParametersAreNonnullByDefault;
