+++ src/main/java/io/github/syst3ms/skriptparser/util/color/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package io.github.syst3ms.skriptparser.util.color;
+
+import javax.annotation.ParametersAreNonnullByDefault;
